fix: Perform book search in componentDidMount
Initially, the book search and resulting setState actions were taking place whenever the user types into the search bar (so after the query change triggered a setState). This led to an issue where the app would try to do the book search and render the results when the SearchBooks component was no longer mounted. Moving the searchBooks() function into componentDidMount() prevents attempting to perform and render the search results when the component is no longer mounted due to the query changing too quickly.