feat: Make dropdowns load while finding shelves
Change the book-shelf-dropdown icon to the loading gif while waiting to figure out which shelf each book is on. Once this information is returned, then hide the loading gif and display the book shelf icons. This was done in order to give the user a visual cue that the page is still loading even though the book covers have already rendered.