fix: Fix memory leaks by managing isMounted state
There was a recurring error where you can't call setState on an unmounted component, which would lead to potential memory leaks. This error was resolved by manually figuring when the component is mounted and is no longer mounted (by setting the _isMounted state within the componentDidMount and componentWillUnmount functions). Then, setState would only be called if the _isMounted was true.